requirements.txt
fer
pydub
datetime
numpy
ffmpeg-python
pathlib
matplotlib
tqdm
opencv-python
deepface
